bad-bits: refactor compare in SQL
