+++ bad-bits/bin/scheduler.js
@@ -4,10 +4,8 @@ export default {
   async scheduled(_controller, env, _ctx) {
     console.log('Running scheduled badbits update...')
     try {
-      const result = await fetchAndStoreBadBits(env)
-      console.log(
-        `Updated badbits denylist: ${result.added} added, ${result.removed} removed`,
-      )
+      await fetchAndStoreBadBits(env)
+      console.log('Updated badbits denylist')
     } catch (error) {
       console.error('Failed to update badbits denylist:', error)
     }
